LUN-208: rename import, move component, fix test
