+++ src/routes/home/__tests__/HomeScreen.spec.tsx
@@ -20,7 +20,7 @@ jest.mock('../../../hooks/useReadCustomDisciplines')
 jest.mock('../../../hooks/useReadSelectedProfessions')
 jest.mock('../../../hooks/useLoadDisciplines')
 jest.mock('../../../hooks/useLoadDiscipline')
-jest.mock('../../../components/HomeScreenHeader', () => {
+jest.mock('../components/HomeScreenHeader', () => {
   const Text = require('react-native').Text
   return () => <Text>HeaderWithMenu</Text>
 })
